refactor: define endpoint for svg thumbnail in a dedicated jsfile
