+++ api/src/endpoints/index.js
@@ -3,6 +3,7 @@ import neo4jRoutes from 'endpoints/neo4j';
 import repoRoutes from 'endpoints/repository';
 import hpaRoutes from 'endpoints/hpaRna';
 import dataOverlayRoutes from 'endpoints/dataOverlay';
+import svgRoutes from 'endpoints/svgThumbnail';
 import swaggerRoutes from 'endpoints/swagger';
 
 const router = express.Router();
@@ -11,6 +12,7 @@ router.use(neo4jRoutes);
 router.use('/repository', repoRoutes);
 router.use('/rna', hpaRoutes);
 router.use('/data-overlay', dataOverlayRoutes);
+router.use('/svg-thumbnail', svgRoutes);
 router.use(swaggerRoutes);
 
 export default router;
